Started updating validation exception messages: passing rule ID to the exception constructor.

+++ core2/src/main/java/org/oboparser/obo/GenSym.java
@@ -16,7 +16,7 @@ public class GenSym {
 	
 	public GenSym(String pref) { 
 		prefix = pref; 
-		id = (long)0;
+		id = 0;
 		setDigits(9);
 	}
